+++ src/World.h
@@ -5,24 +5,26 @@
 
 #include <functional>
 
-#include "Simulator/SimulatorManager.h"
-#include "ChunkMap.h"
-#include "WorldStorage/WorldStorage.h"
+#include "Blocks/BroadcastInterface.h"
+#include "Blocks/WorldInterface.h"
 #include "ChunkGeneratorThread.h"
-#include "ChunkSender.h"
+#include "ChunkMap.h"
 #include "Defines.h"
-#include "LightingThread.h"
+#include "EffectID.h"
+#include "Entities/Boat.h"
+#include "Entities/ProjectileEntity.h"
+#include "ForEachChunkProvider.h"
+#include "Generating/ChunkGenerator.h"
 #include "IniFile.h"
 #include "Item.h"
-#include "Mobs/Monster.h"
-#include "Entities/ProjectileEntity.h"
-#include "Entities/Boat.h"
-#include "ForEachChunkProvider.h"
-#include "Scoreboard.h"
+#include "LightingThread.h"
 #include "MapManager.h"
-#include "Blocks/WorldInterface.h"
-#include "Blocks/BroadcastInterface.h"
-#include "EffectID.h"
+#include "Mobs/Monster.h"
+#include "Protocol/ChunkSender.h"
+#include "Scoreboard.h"
+#include "Simulator/SimulatorManager.h"
+#include "Vector3.h"
+#include "WorldStorage/WorldStorage.h"
